+++ src/pages/meal-plan/meal-plan-page.tsx
@@ -1,8 +1,6 @@
 import { useEffect, useState } from "react";
 import HomeHeader from "../../components/header/HomeHeader";
-import TodayMeal from "../../components/meal-plan/TodayMeal";
-import WeekMeal from "../../components/meal-plan/WeekMeal";
-import { useSearchParams } from "react-router-dom";
+import { useNavigate, useSearchParams } from "react-router-dom";
 import selectedSun from "../../assets/icons/sun-selected.svg";
 import unselectedSun from "../../assets/icons/sun-unselected.svg";
 import selectedMoon from "../../assets/icons/moon-selected.svg";
@@ -11,6 +9,17 @@ import useGetTodayMealPlan from "../../hooks/queries/use-get-today-meal-plan";
 import { useFamilyStore } from "../../stores/use-family-store";
 import EmptyState from "../../components/common/EmptyState";
 import { LoadingSpinner } from "../../components/common/LoadingSpinner";
+import IngredientAndRecipe from "../../components/common/IngredientAndRecipe";
+import Button from "../../components/common/Button";
+import ReviewModal from "../../components/meal-plan/ReviewModal";
+import type { TodayMeal } from "../../types/meal-plan";
+import { useProfileStore } from "../../stores/use-profile-store";
+import { getNextMonday, getThisMonday, getWeekOfMonth } from "../../utils/date";
+import useGetWeekMealPlan from "../../hooks/queries/use-get-week-meal";
+import { changeAdditionalProp } from "../../utils/changeAdditionalProp";
+import ListHeader from "../../components/meal-plan/ListHeader";
+import CalendarChipS from "../../components/meal-plan/CalendarChip/CalendarChipS";
+import DateMenuList from "../../components/meal-plan/DateMenuList";
 
 const MealPlanPage = () => {
   const [searchParams] = useSearchParams();
@@ -103,17 +112,187 @@ const MealPlanPage = () => {
                     <EmptyState text={`${todayTab}식단이 생성되지 않았어요.`} />
                   </div>
                 ) : (
-                  <TodayMeal key={todayTab} data={tabData} />
+                  <TodayMealTab key={todayTab} data={tabData} />
                 )}
               </>
             )}
           </>
         )}
-        {tab == "이번주 식단" && <WeekMeal key="THIS" weekType="THIS" />}
-        {tab == "다음주 식단" && <WeekMeal key="NEXT" weekType="NEXT" />}
+        {tab == "이번주 식단" && (
+          <WeekMealTab key="THISWEEK" weekType="THISWEEK" />
+        )}
+        {tab == "다음주 식단" && (
+          <WeekMealTab key="NEXTWEEK" weekType="NEXTWEEK" />
+        )}
       </div>
     </div>
   );
 };
 
 export default MealPlanPage;
+
+// 오늘의 식단 탭
+function TodayMealTab({ data }: { data: TodayMeal }) {
+  const [isDone, setIsDone] = useState(false);
+  const [isOpen, setIsOpen] = useState(false);
+  const handleOpen = () => setIsOpen(false);
+  return (
+    <>
+      <div className="flex justify-between items-center pt-3">
+        {isOpen && (
+          <ReviewModal
+            recipeId={data.id}
+            onClick={handleOpen}
+            type={data.type}
+          />
+        )}
+        <div>
+          {/* <p className="font-normal text-[16px] leading-[1.5]">우유대신,</p> */}
+          {/*현재 부연설명이 들어오고 있지 않아서 주석 처리 */}
+          <p className="font-semibold text-[18px] tracking-[0.01em] text-gray-600">
+            {data.title.split(" ").map((title, idx) => (
+              <span key={idx}>
+                {title} {idx % 2 === 1 && <br />}
+              </span>
+            ))}
+          </p>
+        </div>
+
+        <img
+          src={data.imageUrl}
+          alt={`${data.title} 이미지`}
+          className="w-41 h-29 object-cover rounded-lg"
+        />
+      </div>
+      {!data.isReviewed && (
+        <div className="pt-6">
+          {!isDone ? (
+            <Button
+              type="button"
+              text="식사 완료"
+              onClick={() => setIsDone(true)}
+            />
+          ) : (
+            <Button
+              type="button"
+              text="리뷰 작성"
+              bgColor="white"
+              onClick={() => setIsOpen(true)}
+            />
+          )}
+        </div>
+      )}
+      <div className="pt-11 pb-29">
+        <IngredientAndRecipe
+          ingredients={data.ingredients}
+          step={data.recipeSteps}
+        />
+      </div>
+    </>
+  );
+}
+
+// 이번주/다음주 탭
+function WeekMealTab({ weekType }: { weekType: "THISWEEK" | "NEXTWEEK" }) {
+  const { familyRoomId } = useFamilyStore();
+  const isLeader = useProfileStore().isLeader;
+  const baseDate = new Date();
+  const date =
+    weekType === "THISWEEK" ? getThisMonday(baseDate) : getNextMonday(baseDate); //이번주/다음주 시작 월요일 날짜
+  const { data, isError, isLoading } = useGetWeekMealPlan(familyRoomId!, date);
+
+  const navigate = useNavigate();
+  if (isError) {
+    return (
+      <div className="pt-43">
+        {isLeader ? (
+          <EmptyState
+            text="아직 식단이 생성되지 않았어요"
+            buttonText="식단 생성"
+            onClick={() => navigate(`/meal-plan/create?week=${weekType}`)}
+          />
+        ) : (
+          <EmptyState
+            text="아직 식단이 생성되지 않았어요"
+            primaryText="방장이 식단을 생성하고 있어요"
+          />
+        )}
+      </div>
+    );
+  }
+  if (isLoading) {
+    return (
+      <div className="w-full pt-60">
+        <LoadingSpinner text="식단을 불러오고 있어요..." />
+      </div>
+    );
+  }
+  const weekData = changeAdditionalProp(data?.result.slots || {}, "WEEK");
+  const listHeaderDate =
+    date + "~" + date[5] + date[6] + "." + (Number(date[8] + date[9]) + 6);
+  const { month, weekKor } = getWeekOfMonth(new Date(date));
+
+  const dayKorMap: Record<string, string> = {
+    MONDAY: "월",
+    TUESDAY: "화",
+    WEDNESDAY: "수",
+    THURSDAY: "목",
+    FRIDAY: "금",
+    SATURDAY: "토",
+    SUNDAY: "일",
+  };
+  const dayNames = [
+    "MONDAY",
+    "TUESDAY",
+    "WEDNESDAY",
+    "THURSDAY",
+    "FRIDAY",
+    "SATURDAY",
+    "SUNDAY",
+  ];
+  const todayIdx = new Date().getDay();
+
+  const adjustedIdx =
+    weekType === "THISWEEK" ? (todayIdx === 0 ? 6 : todayIdx - 1) : 0;
+
+  //오늘~앞으로 날
+  const futureDays = dayNames.slice(adjustedIdx).map((day) => ({
+    dayKor: dayKorMap[day],
+    meals: weekData[day] || [],
+  }));
+
+  //지난 날
+  const pastDays = dayNames.slice(0, adjustedIdx).map((day) => ({
+    dayKor: dayKorMap[day],
+    meals: weekData[day] || [],
+  }));
+  return (
+    <>
+      <div className="pt-[33px] pb-[11px]">
+        <ListHeader title={`${month}월 ${weekKor}`} date={listHeaderDate} />
+      </div>
+      <div className="flex flex-col gap-3 pb-30">
+        {futureDays.map((day) => {
+          if (day.meals.length > 0) {
+            return (
+              <div key={day.dayKor} className="flex gap-3">
+                <CalendarChipS text={day.dayKor} type="primary" />
+                <DateMenuList isSelect={true} data={day.meals} />
+              </div>
+            );
+          }
+        })}
+        {pastDays.map((day) => {
+          if (day.meals.length > 0) {
+            return (
+              <div key={day.dayKor} className="flex gap-3">
+                <CalendarChipS text={day.dayKor} type="gray" />
+                <DateMenuList isSelect={false} data={day.meals} />
+              </div>
+            );
+          }
+        })}
+      </div>
+    </>
+  );
+}
